requirements.txt
yfinance>=0.2.31
pyyaml>=6.0
flask>=3.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
